Sync/libxse re fixes 20260925

+++ include/RE/I/InventoryUserUIInterfaceEntry.h
@@ -10,7 +10,7 @@ namespace RE
 	public:
 		// members
 		InventoryInterface::Handle     invHandle;   // 00
-		BSTSmallArray<std::uint8_t, 4> stackIndex;  // 08
+		BSTSmallArray<std::uint16_t, 4> stackIndex;  // 08
 	};
 	static_assert(sizeof(InventoryUserUIInterfaceEntry) == 0x20);
 }
